requirements.txt
Django==1.4.3
Markdown==2.2.1
PyYAML==3.10
South==0.7.6
distribute==0.6.35
django-braces==0.2.3
django-filter==0.5.4
django-tastypie==0.9.11
djangorestframework==2.1.4
mimeparse==0.1.3
nltk==2.0.4
psycopg2==2.4.5
python-dateutil==1.5
requests==0.14.2
simplemediawiki==1.1.1
